AG-11385 Move nearest.ts into util/
This file is generic enough and doesn't need to know about the scene.
