+++ packages/ag-charts-community/src/chart/series/series.ts
@@ -11,13 +11,13 @@ import type { ScaleType } from '../../scale/scale';
 import type { BBox } from '../../scene/bbox';
 import { Group } from '../../scene/group';
 import type { ZIndexSubOrder } from '../../scene/layersManager';
-import { type DistantObject, nearestSquared } from '../../scene/nearest';
 import type { Node } from '../../scene/node';
 import type { Point } from '../../scene/point';
 import type { PlacedLabel, PointLabelDatum } from '../../scene/util/labelPlacement';
 import { createId } from '../../util/id';
 import { jsonDiff } from '../../util/json';
 import { Listeners } from '../../util/listeners';
+import { type DistantObject, nearestSquared } from '../../util/nearest';
 import { clamp } from '../../util/number';
 import { mergeDefaults } from '../../util/object';
 import type { TypedEvent } from '../../util/observable';
