+++ packages/ag-charts-community/src/integrated-charts-scene.ts
@@ -38,8 +38,6 @@ export { Tooltip, toTooltipHtml } from './chart/tooltip/tooltip';
 export type { TooltipMeta } from './chart/tooltip/tooltip';
 export { BBox } from './scene/bbox';
 export { SectorBox } from './scene/sectorBox';
-export type { DistantObject, NearestResult } from './scene/nearest';
-export { nearestSquared, nearestSquaredInContainer } from './scene/nearest';
 export { HdpiCanvas } from './scene/canvas/hdpiCanvas';
 export { Image } from './scene/image';
 export { ExtendedPath2D } from './scene/extendedPath2D';
