+++ packages/ag-charts-community/src/module-support.ts
@@ -6,6 +6,7 @@ export { extractDecoratedProperties, isDecoratedObject, listDecoratedProperties 
 export * from './util/dom';
 export * from './util/deprecation';
 export * from './util/json';
+export * from './util/nearest';
 export * from './util/number';
 export * from './util/object';
 export * from './util/properties';
