+++ packages/ag-charts-community/src/scene/bbox.ts
@@ -1,7 +1,7 @@
 import { type Interpolating, interpolate } from '../util/interpolating';
+import type { DistantObject, NearestResult } from '../util/nearest';
+import { nearestSquared } from '../util/nearest';
 import { clamp } from '../util/number';
-import type { DistantObject, NearestResult } from './nearest';
-import { nearestSquared } from './nearest';
 
 // For small data structs like a bounding box, objects are superior to arrays
 // in terms of performance (by 3-4% in Chrome 71, Safari 12 and by 20% in Firefox 64).
