+++ packages/ag-charts-community/src/scene/shape/path.ts
@@ -1,5 +1,5 @@
+import type { DistantObject } from '../../util//nearest';
 import { ExtendedPath2D } from '../extendedPath2D';
-import type { DistantObject } from '../nearest';
 import type { RenderContext } from '../node';
 import { RedrawType, SceneChangeDetection } from '../node';
 import { Shape } from './shape';
