+++ packages/ag-charts-community/src/util/nearest.ts
